refactor(#55): перенесен весь код из модуля Test-Trader в модуль Core
