+++ core/src/main/java/io/github/ijlijapol/exception/TestOrderPersistenceException.java
@@ -1,4 +1,4 @@
-package io.github.ijlijapol.bybit.exception;
+package io.github.ijlijapol.exception;
 
 public class TestOrderPersistenceException extends RuntimeException {
     public TestOrderPersistenceException(String message, Exception cause) {
