+++ core/src/main/java/io/github/ijlijapol/model/TestOrder.java
@@ -1,5 +1,6 @@
-package io.github.ijlijapol.bybit.model;
+package io.github.ijlijapol.model;
 
+import io.github.ijlijapol.bybit.model.Symbol;
 import io.github.ijlijapol.bybit.model.order.Side;
 import io.github.ijlijapol.bybit.model.order.TradeOrderType;
 import jakarta.persistence.*;
