+++ core/src/main/resources/application.yaml
@@ -1,0 +1,10 @@
+spring:
+  datasource:
+    driverClassName: org.postgresql.Driver
+    url: jdbc:postgresql://localhost:5432/trading
+    username: dbuser
+    password: 12345
+
+logging:
+  level:
+    root: "INFO"
