+++ core/src/main/resources/schema.sql
